// backend-hedera/custodian/libs/shared/src/guardian/enum/grid-type.enum.ts
export enum GridTypeEnum {
    PROJECT_GRID = 'project_grid',
    USER_GRID = 'user_grid',
    ORGANIZATION_GRID = 'organizations_grid',
    INF_GRID = 'inf_grid',
    PDD_GRID = 'pdd_grid',
    VALIDATION_GRID = 'validation_report_grid',
    ACTIVITY_GRID = 'activity_grid',
    MONITORING_GRID = 'monitoring_report_grid',
    VERIFICATION_GRID = 'verification_report_grid',
}